\name{chromPlot-package}
\alias{chromPlot-package}
\docType{package}
\title{Global visualization tool of genomic data}
\description{Package designed to visualize genomic data along the chromosomes,
where the vertical chromosomes are sorted by number, with sex chromosomes at the
 end.}
\details{
\tabular{ll}{
Package: \tab chromPlot\cr
Type: \tab Package\cr
Version: \tab 0.0.1\cr
Date: \tab 2016-01-05\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Author: Ricardo Verdugo and Karen Orostica
Maintainer: Karen Orostica <korostica09@alumnos.utalca.cl>
}
\keyword{chromPlot}

